@@ -132,20 +136,21 @@
     "@babel/core": "7.14.8",
     "@babel/preset-env": "^7.14.8",
     "@babel/runtime": "7.14.8",
-    "@react-native-community/eslint-config": "3.0.0",
-    "babel-jest": "27.0.6",
-    "eslint": "7.32.0",
-    "eslint-plugin-prettier": "^3.4.0",
-    "eslint-plugin-react": "^7.24.0",
-    "eslint-plugin-react-hooks": "^4.2.0",
+    "@react-native-community/eslint-config": "3.0.1",
+    "babel-jest": "27.4.6",
+    "babel-plugin-root-import": "^6.6.0",
+    "eslint": "8.6.0",
+    "eslint-plugin-prettier": "^4.0.0",
+    "eslint-plugin-react": "^7.28.0",
+    "eslint-plugin-react-hooks": "^4.3.0",
     "husky": "^7.0.1",
     "jest": "27.0.6",
     "lint-staged": "^11.1.1",
     "metro-react-native-babel-preset": "0.66.2",
-    "prettier": "^2.3.2",
+    "prettier": "^2.5.1",
     "react-native-debugger-open": "^0.3.25",
     "react-test-renderer": "17.0.2",
-    "reactotron-react-native": "^5.0.0",
+    "reactotron-react-native": "^5.0.1",
     "reactotron-redux": "^3.1.3",
     "reactotron-redux-saga": "^4.2.3",
     "redux-devtools": "^3.7.0"
