@@ -1,18 +1,27 @@
-import { Animated } from "react-native";
-import ViewPagerAndroid from "@react-native-community/viewpager";
-const React = require("react");
+import { Animated } from 'react-native';
+import ViewPagerAndroid from '@react-native-community/viewpager';
+const React = require('react');
 const { Component } = React;
-const { ViewPropTypes } = (ReactNative = require("react-native"));
-const createReactClass = require("create-react-class");
-const PropTypes = require("prop-types");
-const { Dimensions, View, ScrollView, Platform, StyleSheet, InteractionManager } = ReactNative;
-const TimerMixin = require("react-timer-mixin");
-const SceneComponent = require("./SceneComponent");
-const DefaultTabBar = require("./DefaultTabBar");
-const ScrollableTabBar = require("./ScrollableTabBar");
+const { ViewPropTypes } = (ReactNative = require('react-native'));
+const createReactClass = require('create-react-class');
+const PropTypes = require('prop-types');
+const {
+  Dimensions,
+  View,
+  ScrollView,
+  Platform,
+  StyleSheet,
+  InteractionManager,
+} = ReactNative;
+const TimerMixin = require('react-timer-mixin');
+const SceneComponent = require('./SceneComponent');
+const DefaultTabBar = require('./DefaultTabBar');
+const ScrollableTabBar = require('./ScrollableTabBar');
 
 const AnimatedViewPagerAndroid =
-  Platform.OS === "android" ? Animated.createAnimatedComponent(ViewPagerAndroid) : undefined;
+  Platform.OS === 'android'
+    ? Animated.createAnimatedComponent(ViewPagerAndroid)
+    : undefined;
 
 const ScrollableTabView = createReactClass({
   mixins: [TimerMixin],
@@ -23,7 +32,12 @@ const ScrollableTabView = createReactClass({
   scrollOnMountCalled: false,
 
   propTypes: {
-    tabBarPosition: PropTypes.oneOf(["top", "bottom", "overlayTop", "overlayBottom"]),
+    tabBarPosition: PropTypes.oneOf([
+      'top',
+      'bottom',
+      'overlayTop',
+      'overlayBottom',
+    ]),
     initialPage: PropTypes.number,
     page: PropTypes.number,
     onChangeTab: PropTypes.func,
@@ -38,11 +52,11 @@ const ScrollableTabView = createReactClass({
 
   getDefaultProps() {
     return {
-      tabBarPosition: "top",
+      tabBarPosition: 'top',
       initialPage: 0,
       page: -1,
-      onChangeTab: () => {},
-      onScroll: () => {},
+      onChangeTab: () => { },
+      onScroll: () => { },
       contentProps: {},
       scrollWithoutAnimation: false,
       locked: false,
@@ -51,13 +65,13 @@ const ScrollableTabView = createReactClass({
   },
 
   getInitialState() {
-    const containerWidth = Dimensions.get("window").width;
+    const containerWidth = Dimensions.get('window').width;
     let scrollValue;
     let scrollXIOS;
     let positionAndroid;
     let offsetAndroid;
 
-    if (Platform.OS === "ios") {
+    if (Platform.OS === 'ios') {
       scrollXIOS = new Animated.Value(this.props.initialPage * containerWidth);
       const containerWidthAnimatedValue = new Animated.Value(containerWidth);
       // Need to call __makeNative manually to avoid a native animated bug. See
@@ -66,7 +80,9 @@ const ScrollableTabView = createReactClass({
       scrollValue = Animated.divide(scrollXIOS, containerWidthAnimatedValue);
 
       const callListeners = this._polyfillAnimatedValue(scrollValue);
-      scrollXIOS.addListener(({ value }) => callListeners(value / this.state.containerWidth));
+      scrollXIOS.addListener(({ value }) =>
+        callListeners(value / this.state.containerWidth),
+      );
     } else {
       positionAndroid = new Animated.Value(this.props.initialPage);
       offsetAndroid = new Animated.Value(0);
@@ -98,7 +114,10 @@ const ScrollableTabView = createReactClass({
 
   componentWillReceiveProps(props) {
     if (props.children !== this.props.children) {
-      this.updateSceneKeys({ page: this.state.currentPage, children: props.children });
+      this.updateSceneKeys({
+        page: this.state.currentPage,
+        children: props.children,
+      });
     }
 
     if (props.page >= 0 && props.page !== this.state.currentPage) {
@@ -107,7 +126,7 @@ const ScrollableTabView = createReactClass({
   },
 
   componentWillUnmount() {
-    if (Platform.OS === "ios") {
+    if (Platform.OS === 'ios') {
       this.state.scrollXIOS.removeAllListeners();
     } else {
       this.state.positionAndroid.removeAllListeners();
@@ -116,19 +135,21 @@ const ScrollableTabView = createReactClass({
   },
 
   goToPage(pageNumber) {
-    if (Platform.OS === "ios") {
+    if (Platform.OS === 'ios') {
       const offset = pageNumber * this.state.containerWidth;
       if (this.scrollView) {
-        this.scrollView
-          .getNode()
-          .scrollTo({ x: offset, y: 0, animated: !this.props.scrollWithoutAnimation });
+        this.scrollView.scrollTo({
+          x: offset,
+          y: 0,
+          animated: !this.props.scrollWithoutAnimation,
+        });
       }
     } else {
       if (this.scrollView) {
         if (this.props.scrollWithoutAnimation) {
-          this.scrollView.getNode().setPageWithoutAnimation(pageNumber);
+          this.scrollView?.setPageWithoutAnimation(pageNumber);
         } else {
-          this.scrollView.getNode().setPage(pageNumber);
+          this.scrollView?.setPage(pageNumber);
         }
       }
     }
@@ -150,7 +171,11 @@ const ScrollableTabView = createReactClass({
     }
   },
 
-  updateSceneKeys({ page, children = this.props.children, callback = () => {} }) {
+  updateSceneKeys({
+    page,
+    children = this.props.children,
+    callback = () => { },
+  }) {
     let newKeys = this.newSceneKeys({
       previousKeys: this.state.sceneKeys,
       currentPage: page,
@@ -159,11 +184,18 @@ const ScrollableTabView = createReactClass({
     this.setState({ currentPage: page, sceneKeys: newKeys }, callback);
   },
 
-  newSceneKeys({ previousKeys = [], currentPage = 0, children = this.props.children }) {
+  newSceneKeys({
+    previousKeys = [],
+    currentPage = 0,
+    children = this.props.children,
+  }) {
     let newKeys = [];
     this._children(children).forEach((child, idx) => {
       let key = this._makeSceneKey(child, idx);
-      if (this._keyExists(previousKeys, key) || this._shouldRenderSceneKey(idx, currentPage)) {
+      if (
+        this._keyExists(previousKeys, key) ||
+        this._shouldRenderSceneKey(idx, currentPage)
+      ) {
         newKeys.push(key);
       }
     });
@@ -175,11 +207,11 @@ const ScrollableTabView = createReactClass({
   // to add a listener to `scrollValue`. See https://github.com/facebook/react-native/pull/12620.
   _polyfillAnimatedValue(animatedValue) {
     const listeners = new Set();
-    const addListener = (listener) => {
+    const addListener = listener => {
       listeners.add(listener);
     };
 
-    const removeListener = (listener) => {
+    const removeListener = listener => {
       listeners.delete(listener);
     };
 
@@ -191,32 +223,37 @@ const ScrollableTabView = createReactClass({
     animatedValue.removeListener = removeListener;
     animatedValue.removeAllListeners = removeAllListeners;
 
-    return (value) => listeners.forEach((listener) => listener({ value }));
+    return value => listeners.forEach(listener => listener({ value }));
   },
 
   _shouldRenderSceneKey(idx, currentPageKey) {
     let numOfSibling = this.props.prerenderingSiblingsNumber;
-    return idx < currentPageKey + numOfSibling + 1 && idx > currentPageKey - numOfSibling - 1;
+    return (
+      idx < currentPageKey + numOfSibling + 1 &&
+      idx > currentPageKey - numOfSibling - 1
+    );
   },
 
   _keyExists(sceneKeys, key) {
-    return sceneKeys.find((sceneKey) => key === sceneKey);
+    return sceneKeys.find(sceneKey => key === sceneKey);
   },
 
   _makeSceneKey(child, idx) {
-    return child.props.tabLabel + "_" + idx;
+    return child.props.tabLabel + '_' + idx;
   },
 
   renderScrollableContent() {
-    if (Platform.OS === "ios") {
+    if (Platform.OS === 'ios') {
       const scenes = this._composeScenes();
       return (
         <Animated.ScrollView
           horizontal
           pagingEnabled
           automaticallyAdjustContentInsets={false}
-          contentOffset={{ x: this.props.initialPage * this.state.containerWidth }}
-          ref={(scrollView) => {
+          contentOffset={{
+            x: this.props.initialPage * this.state.containerWidth,
+          }}
+          ref={scrollView => {
             this.scrollView = scrollView;
           }}
           onScroll={Animated.event(
@@ -260,7 +297,7 @@ const ScrollableTabView = createReactClass({
               listener: this._onScroll,
             },
           )}
-          ref={(scrollView) => {
+          ref={scrollView => {
             this.scrollView = scrollView;
           }}
           {...this.props.contentProps}>
@@ -276,7 +313,10 @@ const ScrollableTabView = createReactClass({
       return (
         <SceneComponent
           key={child.key}
-          shouldUpdated={this._shouldRenderSceneKey(idx, this.state.currentPage)}
+          shouldUpdated={this._shouldRenderSceneKey(
+            idx,
+            this.state.currentPage,
+          )}
           style={{ width: this.state.containerWidth }}>
           {this._keyExists(this.state.sceneKeys, key) ? (
             child
@@ -298,7 +338,7 @@ const ScrollableTabView = createReactClass({
 
   _updateSelectedPage(nextPage) {
     let localNextPage = nextPage;
-    if (typeof localNextPage === "object") {
+    if (typeof localNextPage === 'object') {
       localNextPage = nextPage.nativeEvent.position;
     }
 
@@ -318,7 +358,7 @@ const ScrollableTabView = createReactClass({
   },
 
   _onScroll(e) {
-    if (Platform.OS === "ios") {
+    if (Platform.OS === 'ios') {
       const offsetX = e.nativeEvent.contentOffset.x;
       if (offsetX === 0 && !this.scrollOnMountCalled) {
         this.scrollOnMountCalled = true;
@@ -334,16 +374,23 @@ const ScrollableTabView = createReactClass({
   _handleLayout(e) {
     const { width } = e.nativeEvent.layout;
 
-    if (!width || width <= 0 || Math.round(width) === Math.round(this.state.containerWidth)) {
+    if (
+      !width ||
+      width <= 0 ||
+      Math.round(width) === Math.round(this.state.containerWidth)
+    ) {
       return;
     }
 
-    if (Platform.OS === "ios") {
+    if (Platform.OS === 'ios') {
       const containerWidthAnimatedValue = new Animated.Value(width);
       // Need to call __makeNative manually to avoid a native animated bug. See
       // https://github.com/facebook/react-native/pull/14435
       containerWidthAnimatedValue.__makeNative();
-      scrollValue = Animated.divide(this.state.scrollXIOS, containerWidthAnimatedValue);
+      scrollValue = Animated.divide(
+        this.state.scrollXIOS,
+        containerWidthAnimatedValue,
+      );
       this.setState({ containerWidth: width, scrollValue });
     } else {
       this.setState({ containerWidth: width });
@@ -354,15 +401,16 @@ const ScrollableTabView = createReactClass({
   },
 
   _children(children = this.props.children) {
-    return React.Children.map(children, (child) => child);
+    return React.Children.map(children, child => child);
   },
 
   render() {
     let overlayTabs =
-      this.props.tabBarPosition === "overlayTop" || this.props.tabBarPosition === "overlayBottom";
+      this.props.tabBarPosition === 'overlayTop' ||
+      this.props.tabBarPosition === 'overlayBottom';
     let tabBarProps = {
       goToPage: this.goToPage,
-      tabs: this._children().map((child) => child.props.tabLabel),
+      tabs: this._children().map(child => child.props.tabLabel),
       activeTab: this.state.currentPage,
       scrollValue: this.state.scrollValue,
       containerWidth: this.state.containerWidth,
@@ -385,18 +433,21 @@ const ScrollableTabView = createReactClass({
     }
     if (overlayTabs) {
       tabBarProps.style = {
-        position: "absolute",
+        position: 'absolute',
         left: 0,
         right: 0,
-        [this.props.tabBarPosition === "overlayTop" ? "top" : "bottom"]: 0,
+        [this.props.tabBarPosition === 'overlayTop' ? 'top' : 'bottom']: 0,
       };
     }
 
     return (
-      <View style={[styles.container, this.props.style]} onLayout={this._handleLayout}>
-        {this.props.tabBarPosition === "top" && this.renderTabBar(tabBarProps)}
+      <View
+        style={[styles.container, this.props.style]}
+        onLayout={this._handleLayout}>
+        {this.props.tabBarPosition === 'top' && this.renderTabBar(tabBarProps)}
         {this.renderScrollableContent()}
-        {(this.props.tabBarPosition === "bottom" || overlayTabs) && this.renderTabBar(tabBarProps)}
+        {(this.props.tabBarPosition === 'bottom' || overlayTabs) &&
+          this.renderTabBar(tabBarProps)}
       </View>
     );
   },
