@@ -1,24 +1,27 @@
 /**
  * @format
  */
+import { NavigationContainer } from '@react-navigation/native';
+import { createStackNavigator } from '@react-navigation/stack';
+import { AxiosProvider } from '@shared/providers';
+import { Loading, RootComponent } from 'components';
+import * as React from 'react';
+import { StatusBar } from 'react-native';
+import 'react-native-gesture-handler';
+import { KeyboardAwareScrollView } from 'react-native-keyboard-aware-scroll-view';
+import SplashScreen from 'react-native-splash-screen';
+import { Provider } from 'react-redux';
+import { PersistGate } from 'redux-persist/integration/react';
+import { AuthStack, MainStack } from './navigations';
+import { isMountedRef, navigationRef } from './navigations/RootNavigation';
+import configureStore from './redux/store';
 
-import { NavigationContainer } from "@react-navigation/native";
-import { createStackNavigator } from "@react-navigation/stack";
-import { Loading, RootComponent } from "components";
-import { isMountedRef, navigationRef } from "./navigations/RootNavigation";
-import * as React from "react";
-import { StatusBar } from "react-native";
-import codePush from "react-native-code-push";
-import "react-native-gesture-handler";
-import { KeyboardAwareScrollView } from "react-native-keyboard-aware-scroll-view";
-import SplashScreen from "react-native-splash-screen";
-import { Provider } from "react-redux";
-import { PersistGate } from "redux-persist/integration/react";
-import { AuthStack, MainStack } from "./navigations";
-import { persistor, store } from "./redux/store";
+const { persistor, store } = configureStore();
 
 if (__DEV__) {
-  import("../ReactotronConfig.js").then(() => console.log("Reactotron Configured"));
+  import('../ReactotronConfig.js').then(() =>
+    console.log('Reactotron Configured'),
+  );
 }
 
 const Stack = createStackNavigator();
@@ -39,10 +42,10 @@ function SwitchNavigator() {
   );
 }
 
-const App: () => React$Node = () => {
+export default App = () => {
   const [isShowKeyBoard, setShowKeyBoard] = React.useState(false);
 
-  const handleKeyBoardShow = (e) => setShowKeyBoard(true);
+  const handleKeyBoardShow = e => setShowKeyBoard(true);
 
   const handleKeyBoardHide = () => setShowKeyBoard(false);
 
@@ -62,19 +65,22 @@ const App: () => React$Node = () => {
         contentContainerStyle={{ flex: 1 }}
         onKeyboardWillShow={handleKeyBoardShow}
         onKeyboardWillHide={handleKeyBoardHide}>
-        <StatusBar backgroundColor="transparent" translucent={true} barStyle="dark-content" />
+        <StatusBar
+          backgroundColor="transparent"
+          translucent={true}
+          barStyle="dark-content"
+        />
         <Provider store={store}>
           <PersistGate loading={null} persistor={persistor}>
-            <NavigationContainer ref={navigationRef}>
-              <SwitchNavigator />
-              <Loading />
-            </NavigationContainer>
+            <AxiosProvider>
+              <NavigationContainer ref={navigationRef}>
+                <SwitchNavigator />
+                <Loading />
+              </NavigationContainer>
+            </AxiosProvider>
           </PersistGate>
         </Provider>
       </KeyboardAwareScrollView>
     </React.Fragment>
   );
 };
-
-let codePushOptions = { checkFrequency: codePush.CheckFrequency.MANUAL };
-export default codePush(codePushOptions)(App);
