@@ -1,10 +1,10 @@
-import React from "react";
-import { View, StyleSheet, TextInput } from "react-native";
-import { TextInputMask } from "react-native-masked-text";
-import { scaleSize } from "utils";
-import Label from "./Label";
-import Text from "../Text";
-import Configs from "configs";
+import React from 'react';
+import { View, StyleSheet, TextInput } from 'react-native';
+import { TextInputMask } from 'react-native-masked-text';
+import { scaleSize } from 'utils';
+import Label from './Label';
+import Text from '../Text';
+import Configs from '@src/configs';
 const {
   COLORS: { INPUT, RED },
 } = Configs;
@@ -23,7 +23,7 @@ export default function Input(props) {
     styleTextInput = {},
     styleLabel = {},
     contentContainerInput = {
-      justifyContent: "space-between",
+      justifyContent: 'space-between',
       borderBottomWidth: 1,
     },
     ...anyProps
@@ -31,9 +31,13 @@ export default function Input(props) {
   const width_input = width ? width : 350;
   const height_input = height ? height : 61;
   const showError = touched && error;
-  const borderBottomColor = showError ? RED : "#eeeeee";
+  const borderBottomColor = showError ? RED : '#eeeeee';
   return (
-    <View style={{ height: scaleSize(height_input + 30), width: scaleSize(width_input) }}>
+    <View
+      style={{
+        height: scaleSize(height_input + 30),
+        width: scaleSize(width_input),
+      }}>
       <View
         style={[
           {
