@@ -1,19 +1,19 @@
-import { StyleSheet } from "react-native";
-import { scaleSize } from "utils";
-import Configs from "configs";
+import { StyleSheet } from 'react-native';
+import { scaleSize } from 'utils';
+import Configs from '@src/configs';
 const {
   COLORS: { COLOR_MAIN_APP },
 } = Configs;
 const styles = StyleSheet.create({
   container_balance: {
-    alignItems: "center",
-    justifyContent: "space-between",
+    alignItems: 'center',
+    justifyContent: 'space-between',
     paddingHorizontal: scaleSize(15),
-    width: "100%",
+    width: '100%',
     borderRadius: scaleSize(5),
     height: scaleSize(45),
-    backgroundColor: "#FFF8F8",
-    flexDirection: "row",
+    backgroundColor: '#FFF8F8',
+    flexDirection: 'row',
     marginTop: scaleSize(20),
   },
 
@@ -21,14 +21,14 @@ const styles = StyleSheet.create({
     paddingHorizontal: scaleSize(15),
   },
   title_left: {
-    alignSelf: "flex-start",
+    alignSelf: 'flex-start',
     marginVertical: scaleSize(25),
-    fontWeight: "500",
+    fontWeight: '500',
   },
   text_input: {
-    fontWeight: "bold",
+    fontWeight: 'bold',
     fontSize: scaleSize(20),
-    textAlign: "center",
+    textAlign: 'center',
   },
   container_button_submit: {
     // position: "absolute",
@@ -39,8 +39,8 @@ const styles = StyleSheet.create({
   image_creditcard: {
     width: scaleSize(128),
     height: scaleSize(80),
-    resizeMode: "contain",
-    alignSelf: "flex-start",
+    resizeMode: 'contain',
+    alignSelf: 'flex-start',
   },
 });
 
