@@ -1,16 +1,16 @@
-import { StyleSheet } from "react-native";
-import { scaleSize } from "utils";
-import Configs from "configs";
+import { StyleSheet } from 'react-native';
+import { scaleSize } from 'utils';
+import Configs from '@src/configs';
 const { COLORS } = Configs;
 const styles = StyleSheet.create({
   tabStyle: {
     marginTop: scaleSize(20),
-    alignItems: "center",
+    alignItems: 'center',
     paddingHorizontal: scaleSize(16),
   },
   title: {
-    alignSelf: "flex-start",
-    alignItems: "center",
+    alignSelf: 'flex-start',
+    alignItems: 'center',
     marginVertical: scaleSize(25),
   },
 });
