@@ -1,18 +1,26 @@
-import React from "react";
-import { View, ScrollView, RefreshControl } from "react-native";
-import { Header, StatusBar, ItemCardPlaceHolder, Text, FocusAwareStatusBar } from "components";
-import styles from "./styles";
-import { Store, Stylist } from "./widget";
-import { scaleHeight, scaleWidth } from "utils";
-import { useSelector, useDispatch } from "react-redux";
-import actions from "@redux/actions";
-import * as RootNavigation from "navigations/RootNavigation";
-import Configs from "configs";
+import React from 'react';
+import { View, ScrollView, RefreshControl } from 'react-native';
+import {
+  Header,
+  StatusBar,
+  ItemCardPlaceHolder,
+  Text,
+  FocusAwareStatusBar,
+} from 'components';
+import styles from './styles';
+import { Store, Stylist } from './widget';
+import { scaleHeight, scaleWidth } from 'utils';
+import { useSelector, useDispatch } from 'react-redux';
+import actions from '@redux/actions';
+import * as RootNavigation from 'navigations/RootNavigation';
+import Configs from '@src/configs';
 
 export default function index(props) {
   const dispatch = useDispatch();
-  const token = useSelector((state) => state.datalocalReducer.token);
-  const placeholders = useSelector((state) => state.datalocalReducer.placeholders);
+  const token = useSelector(state => state.datalocalReducer.token);
+  const placeholders = useSelector(
+    state => state.datalocalReducer.placeholders,
+  );
 
   const [firstLoading, setFirstLoading] = React.useState(false);
   const [isRefresh, setRefresh] = React.useState(false);
@@ -35,27 +43,31 @@ export default function index(props) {
     }, 1000);
   };
 
-  const afterGetFavourite = (status) => {
+  const afterGetFavourite = status => {
     if (status) {
       setFirstLoading(false);
     }
   };
 
   const favourite_store_setting = useSelector(
-    (state) => state.storeReducer.favourite_store_setting,
+    state => state.storeReducer.favourite_store_setting,
   );
-  const staff_favourites = useSelector((state) => state.staffReducer.staff_favourites);
+  const staff_favourites = useSelector(
+    state => state.staffReducer.staff_favourites,
+  );
 
   const onBack = () => {
     RootNavigation.back();
   };
 
   const renderStaffList = () => {
-    return staff_favourites.map((staff) => <Stylist staff={staff} key={staff.staffId} />);
+    return staff_favourites.map(staff => (
+      <Stylist staff={staff} key={staff.staffId} />
+    ));
   };
 
   const renderStores = () => {
-    return favourite_store_setting.map((merchant) => (
+    return favourite_store_setting.map(merchant => (
       <Store key={merchant.merchantId} item={merchant} />
     ));
   };
@@ -94,15 +106,27 @@ export default function index(props) {
 
   return (
     <View style={styles.container}>
-      <FocusAwareStatusBar barStyle="dark-content" backgroundColor="transparent" />
-      <View style={{ backgroundColor: "#f8f8f8" }}>
+      <FocusAwareStatusBar
+        barStyle="dark-content"
+        backgroundColor="transparent"
+      />
+      <View style={{ backgroundColor: '#f8f8f8' }}>
         <StatusBar barStyle="dark-content" />
-        <Header title="Favourites" headerLeft={true} headerRight={false} onBack={onBack} />
+        <Header
+          title="Favourites"
+          headerLeft={true}
+          headerRight={false}
+          onBack={onBack}
+        />
       </View>
 
       <ScrollView
         refreshControl={
-          <RefreshControl refreshing={isRefresh} onRefresh={refreshFavorites} size={30} />
+          <RefreshControl
+            refreshing={isRefresh}
+            onRefresh={refreshFavorites}
+            size={30}
+          />
         }
         bounces={true}>
         <View style={styles.body}>{renderBody()}</View>
@@ -115,7 +139,7 @@ export default function index(props) {
 const Loading = ({ placeholders }) => {
   return placeholders.map((item, index) => (
     <ItemCardPlaceHolder
-      key={index + ""}
+      key={index + ''}
       width={Configs.CARD_WIDTH}
       height={250}
       borderRadius={5}
