@@ -1,11 +1,11 @@
-import { StyleSheet, Dimensions } from "react-native";
-import { scaleSize, statusBarHeight } from "utils";
-import Configs from "configs";
+import { StyleSheet, Dimensions } from 'react-native';
+import { scaleSize, statusBarHeight } from 'utils';
+import Configs from '@src/configs';
 const {
   COLORS: { COLOR_MAIN_APP },
 } = Configs;
 
-const { width } = Dimensions.get("window");
+const { width } = Dimensions.get('window');
 
 const styles = StyleSheet.create({
   default_tab: {
@@ -15,17 +15,17 @@ const styles = StyleSheet.create({
   tabs: {
     width,
     height: scaleSize(65),
-    alignItems: "center",
-    backgroundColor: "#F8F8F8",
+    alignItems: 'center',
+    backgroundColor: '#F8F8F8',
     borderBottomWidth: 1,
-    borderBottomColor: "#eeeeee",
+    borderBottomColor: '#eeeeee',
   },
   wrapper_header: {
-    backgroundColor: "#F8F8F8",
+    backgroundColor: '#F8F8F8',
     height: scaleSize(124 + statusBarHeight()),
-    width: "100%",
+    width: '100%',
 
-    shadowColor: "#000",
+    shadowColor: '#000',
     shadowOffset: {
       width: 0,
       height: 5,
@@ -41,7 +41,7 @@ const styles = StyleSheet.create({
   },
   seperator: {
     height: scaleSize(10),
-    width: "100%",
+    width: '100%',
   },
 });
 
