@@ -1,24 +1,26 @@
-import React from "react";
-import { Image, ScrollView, View } from "react-native";
-import { useDispatch, useSelector } from "react-redux";
+import React from 'react';
+import { Image, ScrollView, View } from 'react-native';
+import { useDispatch, useSelector } from 'react-redux';
 
-import { setDetailMerchant } from "@redux/actions/storeAction";
-import { ItemCardPlaceHolder, Text, ItemCard } from "components";
-import { scaleSize } from "utils";
-import ICONS from "assets";
-import Configs from "configs";
-import actions from "@redux/actions";
+import { setDetailMerchant } from '@redux/actions/storeAction';
+import { ItemCardPlaceHolder, Text, ItemCard } from 'components';
+import { scaleSize } from 'utils';
+import ICONS from 'assets';
+import Configs from '@src/configs';
+import actions from '@redux/actions';
 
 export default function MerchantList(props) {
   const dispatch = useDispatch();
 
-  const placeholders = useSelector((state) => state.datalocalReducer.placeholders);
-  const loading_store = useSelector((state) => state.storeReducer.loading_store);
+  const placeholders = useSelector(
+    state => state.datalocalReducer.placeholders,
+  );
+  const loading_store = useSelector(state => state.storeReducer.loading_store);
 
-  const { store_tab_home = [] } = useSelector((state) => state.storeReducer);
-  const token = useSelector((state) => state.datalocalReducer.token);
+  const { store_tab_home = [] } = useSelector(state => state.storeReducer);
+  const token = useSelector(state => state.datalocalReducer.token);
 
-  const goToStoreDetail = (merchant) => {
+  const goToStoreDetail = merchant => {
     const { merchantId } = merchant;
     dispatch(setDetailMerchant(merchant));
     dispatch(actions.storeAction.getDetailMerchant(merchantId, token));
@@ -29,17 +31,17 @@ export default function MerchantList(props) {
     return (
       <View
         style={{
-          flexDirection: "row",
-          alignItems: "center",
+          flexDirection: 'row',
+          alignItems: 'center',
           marginBottom: scaleSize(25),
         }}>
         <Image
           style={{
             width: scaleSize(16.5),
             height: scaleSize(16.5),
-            resizeMode: "contain",
+            resizeMode: 'contain',
             marginRight: scaleSize(10),
-            tintColor: "#585858",
+            tintColor: '#585858',
           }}
           source={icon}
         />
@@ -52,10 +54,13 @@ export default function MerchantList(props) {
 
   const ItemList = ({ data }) => {
     return (
-      <ScrollView horizontal showsHorizontalScrollIndicator={false} scrollEventThrottle={1}>
+      <ScrollView
+        horizontal
+        showsHorizontalScrollIndicator={false}
+        scrollEventThrottle={1}>
         {data.map((item, index) => (
           <ItemCard
-            key={index + ""}
+            key={index + ''}
             width={Configs.CARD_WIDTH}
             height={scaleSize(265)}
             borderRadius={5}
@@ -70,10 +75,13 @@ export default function MerchantList(props) {
 
   const ItemListHolder = () => {
     return (
-      <ScrollView horizontal showsHorizontalScrollIndicator={false} scrollEventThrottle={1}>
+      <ScrollView
+        horizontal
+        showsHorizontalScrollIndicator={false}
+        scrollEventThrottle={1}>
         {placeholders.map((item, index) => (
           <ItemCardPlaceHolder
-            key={index + ""}
+            key={index + ''}
             width={Configs.CARD_WIDTH}
             height={scaleSize(265)}
             borderRadius={5}
@@ -86,8 +94,9 @@ export default function MerchantList(props) {
 
   return (
     <>
-      <View style={{ paddingBottom: scaleSize(20), paddingLeft: scaleSize(16) }}>
-        <Headers title="Beauty" icon={ICONS["beauty_merchant"]} />
+      <View
+        style={{ paddingBottom: scaleSize(20), paddingLeft: scaleSize(16) }}>
+        <Headers title="Beauty" icon={ICONS.beauty_merchant} />
         {!loading_store && <ItemList data={store_tab_home} />}
         {loading_store && <ItemListHolder />}
       </View>
