@@ -1,13 +1,13 @@
-import { StyleSheet } from "react-native";
-import { scaleSize } from "utils";
-import Configs from "configs";
+import { StyleSheet } from 'react-native';
+import { scaleSize } from 'utils';
+import Configs from '@src/configs';
 const {
   COLORS: { COLOR_MAIN_APP },
 } = Configs;
 const styles = StyleSheet.create({
   content_text: {
-    alignItems: "center",
-    justifyContent: "center",
+    alignItems: 'center',
+    justifyContent: 'center',
     paddingHorizontal: scaleSize(25),
     marginTop: scaleSize(20),
   },
@@ -15,17 +15,17 @@ const styles = StyleSheet.create({
     paddingHorizontal: scaleSize(15),
   },
   text_input: {
-    fontWeight: "bold",
+    fontWeight: 'bold',
     fontSize: scaleSize(20),
-    textAlign: "center",
+    textAlign: 'center',
   },
   container_button_submit: {
-    position: "absolute",
+    position: 'absolute',
   },
   image_creditcard: {
     width: scaleSize(90),
     height: scaleSize(90),
-    resizeMode: "contain",
+    resizeMode: 'contain',
   },
 });
 
